python/requirements.txt
flask
langchain
langchain-core
langchain-google-genai
python-dotenv
